tests/pyproject.toml
[project]
name = "tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "loguru~=0.7.3",
    "psycopg[binary]~=3.2.13",
    "pytest~=8.4.2",
    "pytest-playwright~=0.7.1",
    "pytest-xdist~=3.8.0",
    "toml~=0.10.2",
]
